backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Email validation and utilities
email-validator==2.1.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Environment and configuration
python-dotenv==1.0.0

# Optional: Redis for session storage (if needed later)
# redis==5.0.1

# Optional: Email service (if using SendGrid, Mailgun, etc.)
# sendgrid==6.10.0
# mailgun==0.1.1 